// speedyGo-ui/src/app/services/models/subscription.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import {User} from "./user";

export interface Subscription {
  durationInDays: number;
  endDate: string;
  id?: number;
  isActive: boolean;
  planName: string;
  price: number;
  startDate: string;
  subscriber:User;
}